+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 0.10.0
+
+- Support Custom Shapes, describing composite shapes/text/image concepts inside an `area`
+- `examples/basic_custom_shape.rb` example
+- Support nesting listeners under a Custom Shape that will automatically get added to its `body_root` control
+- Support nesting listeners under a Custom Control that will automatically get added to its `body_root` control
+
 ## 0.9.7
 
 - Scaffold Custom Control Gem via `glimmer scaffold:gem:customcontrol[name,namespace]` (or alias: `glimmer scaffold:gem:cc[name,namespace]`)